+++ helper_test.go
@@ -20,7 +20,7 @@ func TestIsEXE(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.in, func(t *testing.T) {
-			file, err := New(tt.in, nil)
+			file, err := New(tt.in, &Options{})
 			if err != nil {
 				t.Fatalf("New(%s) failed, reason: %v", tt.in, err)
 			}
@@ -50,7 +50,7 @@ func TestIsDLL(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.in, func(t *testing.T) {
-			file, err := New(tt.in, nil)
+			file, err := New(tt.in, &Options{})
 			if err != nil {
 				t.Fatalf("New(%s) failed, reason: %v", tt.in, err)
 			}
@@ -80,7 +80,7 @@ func TestIsDriver(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.in, func(t *testing.T) {
-			file, err := New(tt.in, nil)
+			file, err := New(tt.in, &Options{})
 			if err != nil {
 				t.Fatalf("New(%s) failed, reason: %v", tt.in, err)
 			}