+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- Unit tests for bound imports directory [#18](https://github.com/saferwall/pe/pull/18).
-- Unit tests for security directory [#19](https://github.com/saferwall/pe/pull/19).
+- Unit tests for delay imports directory [#23](https://github.com/saferwall/pe/pull/23).
 - Allow access to the raw certificates content [#22](https://github.com/saferwall/pe/pull/22).
+- Unit tests for security directory [#19](https://github.com/saferwall/pe/pull/19).
+- Unit tests for bound imports directory [#18](https://github.com/saferwall/pe/pull/18).
 
 ## Changed